requirements.txt
google-cloud-storage==2.14.0
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
Pillow==10.4.0
streamlit==1.38.0
PyYAML==6.0.1
requests==2.32.3
fsspec==2024.6.1
gcsfs==2024.6.1
pandas